fixed libs again
