+++ .gitignore
@@ -1,0 +1,7 @@
+/audiofy/cmake-build-debug/audiofy.cbp
+/audiofy/cmake-build-debug/audiofy.exe
+/audiofy/cmake-build-debug/audiofy.ilk
+/audiofy/cmake-build-debug/audiofy.pdb
+/audiofy/cmake-build-debug/audiofy_lib.lib
+/audiofy/cmake-build-debug/CMakeCache.txt
+/audiofy/cmake-build-debug/imgui.ini
