+++ audiofy/CMakeLists.txt
@@ -4,8 +4,15 @@ set(CMAKE_CXX_STANDARD 20)
 add_compile_options(/std:c++latest)
 
 file(GLOB_RECURSE imgui_SRC "src/gui/imgui/*.cpp")
-add_library(imgui_win32_dx11 ${imgui_SRC})
-target_include_directories(imgui_win32_dx11 PUBLIC include/imgui)
+file(GLOB_RECURSE audiolib_SRC "../audiolib/platform/src/*.cpp")
+add_library(imgui_win32_dx11_lib ${imgui_SRC})
+add_library(audiofy_lib ${audiolib_SRC})
+target_include_directories(audiofy_lib PUBLIC ../audiolib/vorbis)
+
+target_link_libraries(audiofy_lib Mfreadwrite.lib Mfplat.lib mfuuid.lib Pathcch.lib
+        Shlwapi.dll)
+
+target_include_directories(imgui_win32_dx11_lib PUBLIC include/imgui)
 
 
 add_executable(audiofy src/Application.cpp src/win32/ay_fileManager.cpp src/audio/AudioWorkspace.cpp src/audio/AudioWorkspace.h
@@ -14,15 +21,17 @@ add_executable(audiofy src/Application.cpp src/win32/ay_fileManager.cpp src/audi
 
 target_include_directories(audiofy PUBLIC include)
 target_include_directories(audiofy PUBLIC include/vorbis)
+target_include_directories(audiofy PUBLIC include/imgui)
 
 target_link_directories(audiofy PUBLIC lib)
-target_link_libraries(audiofy lib_win32 comsuppw)
+target_link_libraries(audiofy comsuppw)
 target_link_libraries(audiofy Mfreadwrite.lib Mfplat.lib mfuuid.lib Pathcch.lib
         Shlwapi.dll opengl32.lib)
 target_link_directories(audiofy PUBLIC lib)
 target_link_directories(audiofy PUBLIC lib/glfw/lib-vc2010-64)
 
-target_link_libraries(audiofy imgui_win32_dx11)
+target_link_libraries(audiofy imgui_win32_dx11_lib)
 target_link_libraries(audiofy SoundTouchDLL_x64)
 target_link_libraries(audiofy D3D11)
+target_link_libraries(audiofy audiofy_lib)
 set_target_properties(audiofy PROPERTIES WIN32_EXECUTABLE TRUE)
