--- audiofy/lib/usynergy/README.txt
@@ -1,8 +1,0 @@
-
-uSynergy client -- Implementation for the embedded Synergy client library
-version 1.0.0, July 7th, 2012
-Copyright (c) 2012 Alex Evans
-
-This is a copy of the files once found at:
-  https://github.com/symless/synergy-core/tree/790d108a56ada9caad8e56ff777d444485a69da9/src/micro
-
