+++ audiofy/src/Application.cpp
@@ -91,18 +91,17 @@ int WinMain(  _In_ HINSTANCE hInstance,
 
 
     auto audioFile = Application::decoder.loadAudioFile(L"allTheTime.mp3");
-    //auto audioFile2 = Application::decoder.loadAudioFile(L"duvet.ogg");
+    auto audioFile2 = Application::decoder.loadAudioFile(L"duvet.ogg");
 
     Application::decoder.decodeAudioFile(audioFile, buffer);
     //Application::decoder.decodeAudioFile(audioFile2, buffer2);
 
-    //auto plot = new WaveformPlot(buffer);
-    //plot->AddBuffer(buffer2);
+    auto plot = new WaveformPlot(buffer);
+    plot->AddBuffer(buffer);
 
     Application::player.playAudioBuffer(buffer);
-
     Application::player.playAudioBuffer(buffer);
-    
+    GuiMain::AddComponent(plot);
     
     while (!done)
     {
