+++ audiolib/platform/src/al_io.h
@@ -8,19 +8,6 @@
 #include<string>
 
 namespace al_io {
-    class OpenFileDialog {
-    public:
-        void showDialog();
-
-    private:
-        std::wstring startPath;
-    };
-    class FileManager {
-    public:
-        static wchar_t* getCurrentExecutablePath();
-        static void setCurrentWorkingDirectory(const wchar_t* path);_
-
-    };
 }
 
 #endif //AUDIOFY_LIB_AL_IO_H
