--- audiolib/platform/src/dummyPlayback.h
@@ -1,8 +1,0 @@
-//
-// Created by Jonathan on 09.11.2021.
-//
-
-#ifndef AUDIOFY_LIB_DUMMYPLAYBACK_H
-#define AUDIOFY_LIB_DUMMYPLAYBACK_H
-
-#endif //AUDIOFY_LIB_DUMMYPLAYBACK_H
